final commit

+++ controllers/videoController.js
@@ -1,5 +1,8 @@
 import routes from "../routes"
 import Video from "../models/Video";
+import Comment from "../models/Comment";
+import cookieParser from "cookie-parser";
+import { comment } from "postcss";
 
 export const home = async (req, res) => {
     try{
@@ -36,9 +39,11 @@ export const postUpload = async (req, res) => {
      const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
-    })
-    console.log(newVideo)
+        description,
+        creator: req.user.id
+    });
+    req.user.videos.push(newVideo);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 export const videoDetail = async(req, res) => {
@@ -46,7 +51,9 @@ export const videoDetail = async(req, res) => {
         params: {id}
     } = req;
     try{
-        const video = await Video.findById(id).populate("creator");
+        const video = await Video.findById(id)
+        .populate("creator")
+        .populate("comments");
         res.render("videoDetail", { pageTitle : video.title, video });
     } catch(error) {
         res.redirect(routes.home);
@@ -59,7 +66,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle : `Edit ${video.title}`, video});
+        if(video.creator.toString() !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle : `Edit ${video.title}`, video});
+        }
     } catch(error){
         res.redirect(routes.home);
     }
@@ -83,10 +94,73 @@ export const deleteVideo = async (req, res) => {
         params: {id}
     } = req;
     try{
-       await Video.findOneAndRemove({_id: id});
+        const video = await Video.findById(id);
+        if(video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id: id});
+        }
     } catch(error) {
         console.log(error);
     }
     res.redirect(routes.home);
-}
+};
+
+// 시청 조회수 Regiseter
+
+export const postRegisterView = async (req, res) => {
+    const {
+        params: {id}
+    } = req;
+    try{
+        const video = await Video.findById(id);
+        video.views += 1;
+        video.save();
+        res.status(200);
+    }catch(error){
+        res.status(400);
+    }finally{
+        res.end();
+    }
+};
+
+// 코멘트 달기
+
+export const postAddComment = async( req, res ) => {
+    const {
+        params: {id},
+        body: { comment },
+        user
+    } = req;
+    try{
+        const video = await Video.findById(id);
+        const newComment = await Comment.create({
+            text: comment,
+            creator: user.id
+        });
+        video.comments.push(newComment._id);
+        video.save();
+    }catch(error){
+        res.status(400);
+    }finally{
+        res.end()
+    }
+};
+
+export const postDeleteComment = async( req, res ) =>{
+    const {
+        params: {id},
+        body: {commentId}
+    } = req; 
+    try{
+        const video = await Video.findById(id);
+        await Comment.findOneAndRemove({_id: commentId});
+        video.comments.findOneAndRemove(commentId);
+        video.save();
+    }catch(error){
+        res.status(400);
+    }finally{
+        res.end();
+    }
+};
